+++ maestro-client/src/main/java/maestro/Driver.kt
@@ -105,6 +105,8 @@ interface Driver {
 
     fun setAndroidChromeDevToolsEnabled(enabled: Boolean) = Unit
 
+    fun sendBroadcast(action: String, receiver: String?, extras: Map<String, Any>?) = Unit
+
     fun queryOnDeviceElements(query: OnDeviceElementQuery): List<TreeNode> {
         return listOf()
     }