+++ maestro-client/src/main/java/maestro/drivers/CdpWebDriver.kt
@@ -248,6 +248,10 @@ class CdpWebDriver(
         stopApp(appId)
     }
 
+    override fun sendBroadcast(action: String, receiver: String?, extras: Map<String, Any>?) {
+        LOGGER.warn("sendBroadcast not available on web")
+    }
+
     override fun contentDescriptor(excludeKeyboardElements: Boolean): TreeNode {
         ensureOpen()