+++ maestro-orchestra/src/main/java/maestro/orchestra/yaml/YamlFluentCommand.kt
@@ -45,6 +45,7 @@ import maestro.orchestra.HideKeyboardCommand
 import maestro.orchestra.InputRandomCommand
 import maestro.orchestra.InputRandomType
 import maestro.orchestra.InputTextCommand
+import maestro.orchestra.SendBroadcastCommand
 import maestro.orchestra.KillAppCommand
 import maestro.orchestra.LaunchAppCommand
 import maestro.orchestra.MaestroCommand
@@ -139,6 +140,7 @@ data class YamlFluentCommand(
     val addMedia: YamlAddMedia? = null,
     val setAirplaneMode: YamlSetAirplaneMode? = null,
     val toggleAirplaneMode: YamlToggleAirplaneMode? = null,
+    val sendBroadcast: YamlSendBroadcast? = null,
     val retry: YamlRetryCommand? = null,
     @JsonIgnore val _location: JsonLocation,
 ) {
@@ -463,6 +465,17 @@ data class YamlFluentCommand(
                     ToggleAirplaneModeCommand(
                         toggleAirplaneMode.label,
                         toggleAirplaneMode.optional
+                    )
+                )
+            )
+
+            sendBroadcast != null -> listOf(
+                MaestroCommand(
+                    SendBroadcastCommand(
+                        action = sendBroadcast.action ?: throw SyntaxError("Broadcast command requires an action"),
+                        receiver = sendBroadcast.receiver,
+                        extras = sendBroadcast.extras,
+                        label = sendBroadcast.label,
                     )
                 )
             )